+++ cog.scm
@@ -1,3 +1,3 @@
 (define package-name 'http2curl)
-(define version "0.2.0")
+(define version "0.3.0")
 (define dependencies '())